requirements.txt
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
eventlet==0.33.2
Flask==2.2.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Panda3D==1.10.12
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
six==1.16.0
urllib3==1.26.12
websocket-client==1.4.2
Werkzeug==2.2.2
numpy==1.21.1
